requirements.txt
requests~=2.26.0
pandas~=1.3.5
matplotlib~=3.5.1
python-binance
numpy~=1.21.5
keras
tensorflow
scikit-learn
sktime
websocket-client